Extract Threading.kt
