+++ compose/ui/ui/src/jsWasmMain/kotlin/androidx/compose/ui/internal/Threading.jsWasm.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2025 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,6 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui
+package androidx.compose.ui.internal
 
-import kotlin.system.getTimeMillis
-import kotlinx.atomicfu.atomic
-
-internal actual fun areObjectsOfSameType(a: Any, b: Any): Boolean {
-    return a::class == b::class
-}
-
-internal actual fun currentTimeMillis(): Long {
-    @Suppress("DEPRECATION") // TODO: Avoid using deprecated function
-    return getTimeMillis()
-}
-
+internal actual fun getCurrentThreadId(): Long = 0
