+++ compose/ui/ui/src/nativeMain/kotlin/androidx/compose/ui/internal/Threading.native.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2025 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,17 +14,13 @@
  * limitations under the License.
  */
 
-package androidx.compose.ui
+package androidx.compose.ui.internal
 
-import kotlin.system.getTimeMillis
 import kotlinx.atomicfu.atomic
 
-internal actual fun areObjectsOfSameType(a: Any, b: Any): Boolean {
-    return a::class == b::class
-}
+private val threadCounter = atomic(0L)
 
-internal actual fun currentTimeMillis(): Long {
-    @Suppress("DEPRECATION") // TODO: Avoid using deprecated function
-    return getTimeMillis()
-}
+@ThreadLocal
+private var threadId: Long = threadCounter.addAndGet(1)
 
+internal actual fun getCurrentThreadId(): Long = threadId
